Enable altitude-based zonal plotting
The zonal plot functions now accept is_pressure=False, in which case
altitude (in km) is used instead of pressure (in hPa). In this case,
either the vertical grid is required (at which point ISA pressure-
altitude is used), or the user must explicitly supply the altitudes
of the cell edges as z_edge.

Signed-off-by: Sebastian D. Eastham <seastham@mit.edu>