Add code to find lat/lon index containing a given lat/lon location
gcgridobj.latlontools.find_idx will robustly find the lat/lon index
containing a specific location, if and only if the grid definition
(given as a vector of edges) is for a regular grid. This code does
allow for half-polar data, and is much more efficient than testing
against grid edge locations.

Signed-off-by: Sebastian D. Eastham <seastham@mit.edu>