Modifications to enable accurate plotting of cubed-sphere data which crosses 180E
An upcoming update to Cartopy (expected in 0.19.0) enables pcolormesh to correctly
display polygons crossing the antimeridian. This removes the need for thresholding
at that location, and means that ALL cubed sphere data is now shown on plots. Users
can recover the old functionality by passing the argument `cs_threshold=5` when
calling `plot_layer` (or the lower-level `plot_cs` function). A side effect of this
fix is that a wrapped set of polgyons is sometimes generated when plotting one of
the cubed-sphere faces. `plottools.set_clim` has been modified to account for this.